Split tech stack content components
